# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TokenInsights REST API
  version: 2.0.0
  description:
    Query dataset-scoped usage and ingest sanitized collector evidence. Personal servers are
    unauthenticated; hosted servers authenticate bearer tokens or read-only browser sessions.
    Servers interpret evidence and never scan host artifacts.
...
paths:
...
  /api/v3/ingestion/batches/{stream}/{batch}:
    get:
      security:
        - bearerAuth: []
        - {}
      operationId: getRawBatchStatusV3
      summary: Read immutable acceptance and current per-item processing status
      tags:
        - Ingestion
      parameters:
        - name: stream
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 256
        - name: batch
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 256
      responses:
        '200':
          description:
            Durably accepted; all items terminal for current processing generation and scope
            revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RawAcceptanceV3'
        '202':
          description: Durably accepted; processing remains asynchronous.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RawAcceptanceV3'
        '503':
          description: Cannot durably accept; retry exact bytes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '404':
          description: No accepted batch with this identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '401':
          description: Hosted authentication required; no acceptance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '403':
          description: Hosted ingestion permission required; browser sessions are read-only.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '429':
          description: Admission limit reached; retry exact bytes after backoff.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
...
  /api/v2/ingestion/batches/{stream}/{batch}:
    get:
      operationId: getRawBatchStatus
      summary: Read immutable acceptance and current per-item processing status
      tags:
        - Ingestion
      parameters:
        - name: stream
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 256
        - name: batch
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 256
      responses:
        '200':
          description:
            Durably accepted; all items terminal for current processing generation and scope
            revision.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RawAcceptance'
        '202':
          description: Durably accepted; processing remains asynchronous.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RawAcceptance'
        '503':
          description: Cannot durably accept; retry exact bytes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
        '404':
          description: No accepted batch with this identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestionError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Hosted token; effective dataset and read/ingest scopes resolved by
        the server.
    browserSession:
      type: apiKey
      in: cookie
      name: tokeninsights_session
      description: Read-only, HttpOnly browser session established by POST /api/v2/auth/session.
...
  schemas:
...
    Count:
      type: integer
      format: int64
      minimum: 0
      maximum: 9007199254740991
    TimestampMs:
      type: integer
      format: int64
      minimum: 0
      maximum: 253402214399999
      description:
        Unix milliseconds through 9999-12-30T23:59:59.999Z. One day of headroom preserves calendar
        query support after timezone conversion. Source units are never guessed or rescaled.
...
    IngestionError:
      type: object
      additionalProperties: false
      required:
        - code
        - stage
      properties:
        code:
          type: string
          maxLength: 256
          description: At most 256 UTF-8 bytes; no control characters.
          minLength: 1
        stage:
          type: string
          maxLength: 256
          description: At most 256 UTF-8 bytes; no control characters.
          minLength: 1
        batchId:
          type: string
          maxLength: 256
          description: At most 256 UTF-8 bytes; no control characters.
          minLength: 1
        factId:
          type: string
          maxLength: 256
          description: At most 256 UTF-8 bytes; no control characters.
          minLength: 1
      description: Safe diagnostic codes and correlation IDs; request content is never echoed.
...
    RawReceipt:
      type: object
      additionalProperties: false
      required:
        - databaseId
        - datasetId
        - streamId
        - batchId
        - requestHash
        - fromSequence
        - toSequence
        - accepted
        - acceptedAtMs
        - inputRevision
      properties:
        databaseId:
          type: string
          minLength: 1
          maxLength: 256
        datasetId:
          type: string
          enum:
            - default
        streamId:
          type: string
          minLength: 1
          maxLength: 256
        batchId:
          type: string
          minLength: 1
          maxLength: 256
        requestHash:
          type: string
          minLength: 1
          maxLength: 256
        fromSequence:
          $ref: '#/components/schemas/Count'
        toSequence:
          $ref: '#/components/schemas/Count'
        accepted:
          $ref: '#/components/schemas/Count'
        acceptedAtMs:
          $ref: '#/components/schemas/TimestampMs'
        inputRevision:
          $ref: '#/components/schemas/Count'
    EvidenceOutcome:
      type: object
      additionalProperties: false
      required:
        - evidenceId
        - disposition
        - generation
        - inputRevision
      properties:
        evidenceId:
          type: string
          minLength: 1
          maxLength: 256
        disposition:
          type: string
          enum:
            - pending
            - context
            - processed
            - duplicate
            - superseded
            - ambiguous
        code:
          type: string
        factId:
          type: string
        generation:
          $ref: '#/components/schemas/Count'
        inputRevision:
          $ref: '#/components/schemas/Count'
    ProcessingStatus:
      type: object
      additionalProperties: false
      required:
        - generation
        - inputRevision
        - pending
        - items
      properties:
        generation:
          $ref: '#/components/schemas/Count'
        inputRevision:
          $ref: '#/components/schemas/Count'
        pending:
          $ref: '#/components/schemas/Count'
        items:
          type: array
          items:
            $ref: '#/components/schemas/EvidenceOutcome'
    RawAcceptance:
      type: object
      additionalProperties: false
      required:
        - receipt
        - processing
      properties:
        receipt:
          $ref: '#/components/schemas/RawReceipt'
        processing:
          $ref: '#/components/schemas/ProcessingStatus'
...
    RawReceiptV3:
      type: object
      additionalProperties: false
      required:
        - databaseId
        - datasetId
        - streamId
        - batchId
        - requestHash
        - fromSequence
        - toSequence
        - accepted
        - acceptedAtMs
        - inputRevision
      properties:
        databaseId:
          type: string
          minLength: 1
          maxLength: 256
        datasetId:
          type: string
          minLength: 1
          maxLength: 256
        streamId:
          type: string
          minLength: 1
          maxLength: 256
        batchId:
          type: string
          minLength: 1
          maxLength: 256
        requestHash:
          type: string
          minLength: 1
          maxLength: 256
        fromSequence:
          $ref: '#/components/schemas/Count'
        toSequence:
          $ref: '#/components/schemas/Count'
        accepted:
          $ref: '#/components/schemas/Count'
        acceptedAtMs:
          $ref: '#/components/schemas/TimestampMs'
        inputRevision:
          $ref: '#/components/schemas/Count'
    RawAcceptanceV3:
      type: object
      additionalProperties: false
      required:
        - receipt
        - processing
      properties:
        receipt:
          $ref: '#/components/schemas/RawReceiptV3'
        processing:
          $ref: '#/components/schemas/ProcessingStatus'
...